split the browser compiler from the server
