+++ .eslintignore
@@ -1,4 +1,5 @@
 lib/browser/wrap
+lib/browser/compiler/wrap/
 test/polyfills
 test/compiler/js
 test/specs/compiler/fixtures
