+++ lib/browser/compiler/index.js
@@ -1,0 +1,5 @@
+import './wrap/prefix'
+import './parsers'
+import '../../helpers/compiler-shared'
+import './core'
+import './wrap/suffix'
