+++ lib/browser/compiler/wrap/prefix.js
@@ -1,0 +1,1 @@
+;(function(window) {
