+++ lib/browser/compiler/wrap/suffix.js
@@ -1,0 +1,1 @@
+})(this)
