+++ lib/node/parsers.js
@@ -1,0 +1,24 @@
+module.exports = {
+  html: {
+    jade: function(html) {
+      return require('jade').render(html, {pretty: true})
+    }
+  },
+  js: {
+    plainjs: function(js) {
+      return js
+    },
+    livescript: function(js) {
+      return require('LiveScript').compile(js, { bare: true, header: false })
+    },
+    typescript: function(js) {
+      return require('typescript-simple')(js)
+    },
+    es6: function(js) {
+      return require('babel').transform(js, { blacklist: ['useStrict'] }).code
+    },
+    coffee: function(js) {
+      return require('coffee-script').compile(js, { bare: true })
+    }
+  }
+}
