+++ lib/riot+compiler.js
@@ -1,2 +1,2 @@
 import 'riot'
-import 'browser/compiler'
+import 'browser/compiler/index'
